requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==5.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.11.1
google-auth==2.6.6
idna==3.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toposort==1.7
urllib3==1.26.9
